codacy recommendations
